RegisterUserNameView
+Major Update on SRC
-DROP DUPLICATES FROM USERNAMEVIEW
