+++ watering/Presentation/RegisterUser/View/Components/UserNameView.swift
@@ -1,66 +1,67 @@
+////
+////  NameUser.swift
+////  Watering
+////
+////  Created by Narely Lima on 29/09/22.
+////
 //
-//  NameUser.swift
-//  Watering
+//import Foundation
+//import SwiftUI
 //
-//  Created by Narely Lima on 29/09/22.
+//struct UserNameView : View {
+//    @State var personName: String = ""
+////    let notification = NotificationController()
+//    var body: some View {
+//        ZStack{
+//            Image("Flower")
+//                .resizable()
+//                .scaledToFit()
+//                .frame(alignment: .trailing)
+//            VStack {
+//                Spacer().frame(height: 50)
+//                question
+//                digiteAqui
+//                Divider()
+//                    .background(ThemeEnum.primary)
+//                    .frame(width: 300)
+//                Spacer().frame(height: 50)
+//                navigation
+//            }
+//        }
+//    }
+//    var question: some View {
+//        Text("What is your name?")
+//            .foregroundColor(ThemeEnum.primary)
+//            .font(.system(size: 22, design: .rounded))
 //
-
-import Foundation
-import SwiftUI
-
-struct UserNameView : View {
-    @State var personName: String = ""
-//    let notification = NotificationController()
-    var body: some View {
-        ZStack{
-            ThemeEnum.flower
-                .resizable()
-                .scaledToFit()
-            VStack {
-                Spacer().frame(height: 50)
-                question
-                digiteAqui
-                Divider()
-                    .background(ThemeEnum.primary)
-                    .frame(width: 300)
-                Spacer().frame(height: 50)
-                navigation
-            }
-        }
-    }
-    var question: some View {
-        Text("What is your name?")
-            .foregroundColor(ThemeEnum.primary)
-            .font(.system(size: 22, design: .rounded))
-
-    }
-    var digiteAqui: some View{
-        TextField("Type Here", text: $personName)
-            .multilineTextAlignment(.center)
-            .foregroundColor(.gray)
-            .font(.system(size: 20, design: .rounded))
-            .padding()
-            .onChange(of: personName) { newValue in
-                saveName()
-            }
-    }
-    var navigation: some View {
-        NavigationLink(destination: RegisterPlantNameView().navigationBarHidden(true), label: {
-            Text("Next")
-                .foregroundColor(ThemeEnum.primary)
-                .padding()
-                .frame(width: 200, height: 50)
-                .font(.system(size: 20, design: .rounded))
-                .overlay(RoundedRectangle(cornerRadius: 15)
-                    .stroke(ThemeEnum.primary, lineWidth: 2))
-                .padding()
-                .cornerRadius(20)
-        }
-        )
-    }
-    func saveName() {
-        UserDefaults.standard.setPersonName(value: personName)
-//        notification.requestPermission()
-    }
-
-}
+//    }
+//    var digiteAqui: some View{
+//        TextField("Type Here", text: $personName)
+//            .multilineTextAlignment(.center)
+//            .foregroundColor(.gray)
+//            .font(.system(size: 20, design: .rounded))
+//            .padding()
+//            .onChange(of: personName) { newValue in
+//                saveName()
+//            }
+//    }
+//    var navigation: some View {
+//        NavigationLink(destination: RegisterPlantNameView().navigationBarHidden(true), label: {
+//            Text("Next")
+//                .foregroundColor(ThemeEnum.primary)
+//                .padding()
+//                .frame(width: 200, height: 50)
+//                .font(.system(size: 20, design: .rounded))
+//                .overlay(RoundedRectangle(cornerRadius: 15)
+//                    .stroke(ThemeEnum.primary, lineWidth: 2))
+//                .padding()
+//                .cornerRadius(20)
+//        }
+//        )
+//    }
+//    func saveName() {
+//        UserDefaults.standard.setPersonName(value: personName)
+////        notification.requestPermission()
+//    }
+//
+//}
